+++ tests/mysql/mod.rs
@@ -1,4 +1,4 @@
-use sea_query::{tests_cfg::*, *};
+use sea_query::{extension::mysql::*, tests_cfg::*, *};
 
 mod foreign_key;
 mod index;